Relax is_handleable check
By now, an increased number of devices can be "handled" by udiskie,
compared to the original version. Rather than being only able to mount
the following methods have been added:

- unlock LUKS slave partitions
- lock LUKS holder partitions
- detach external drives
- eject external media

Checking whether any of the actions can be performed on a given device
is rather tedious. For now, the is_handleable check is thus relaxed to
check only if the device is external.