Remove obsolete exception checks and default icon fallback code
The `gtk.image_new_from_icon_name` function does not raise an exception
in case it does not find the requested icon. Instead, it quietly serves
a default icon indicating to the user that the icon is not available.
This means that the execution branch that returns the default fallback
icon is not available as desired.

It should be noted that, since installing the icons in `setup.py`, it
should be reasonably safe to assume the icons are available.