Revert "Make udiskie a namespace package"
This reverts commit dc9b5d5733e50d807b2fa828f314e8254dba91a5.

The initial thought was to allow tray plugins and similar to share the
same namespace. However, I don't think anymore that udiskie is suited to
be a namespace package. Furthermore, the tray icon code has been merged
into the main repository. Plugins should really live in their own
namespace or be integrated.