Catch exceptions when trying to show notifications
..and simply log the error to the console.

This fixes udiskie's behaviour if no notification service is installed:

Previously, in this case an exception was raised when trying to show the
notification. The exception was caught in the main loop causing udiskie
not to crash altogether. With the new 'device_added' notifications this
could happen before trying to automount, hence making udiskie useless.