Remove Method(s)Proxy
These classes will become obsolete in the async udiskie of the future:
On most devices no methods will ever be executed (except for property
access). Therefore it's unnecessary to preemptively query proxies for
interfaces. Furthermore on-attribute-access querying of interfaces like
was done in the udisks2 API becomes impractical, because attribute is
synchronous while proxy creation is asynchronous.