Fix udisks1 event handling
In udisks1 event handlers there is no way to instantaneously infer the
new device state. The new state be requested via DBus which can lead to
race conditions: the change in the known to the obtained device is not
guaranteed to match the event. Therefore, all events must be handled
equally…