Remove job_failed event from --event-hook
It is not working, and never was.

Resolves: #324

job_failed has additional parameters that the other events do not have.
Because the event_hook command line is the same for all connected
events, it would not be possible to make use of the 'action' and
'message' parameters without breaking the other events. Therefore,
'job_failed' would have to be used without 'action' and 'message',
decreasing its usefulness.

For proper support, we should provide a way to specify hook per event
as requested in #285.